⛲️🛋 Checkpoint ./src/presenters/includes/team-elements.js:271885/360 ./src/presenters/pages/team.js:271885/1294 ./src/components/fields/team-name-input.js:271885/80 ./stories/inputs.js:271885/267
