+++ src/presenters/includes/team-elements.js
@@ -6,14 +6,6 @@ import TooltipContainer from 'Components/tooltips/tooltip-container';
 import Text from 'Components/text/text';
 import Link from 'Components/link';
 
-export const VerifiedBadge = () => {
-  const image = 'https://cdn.glitch.com/55f8497b-3334-43ca-851e-6c9780082244%2Fverified.svg?1501783108220';
-  const tooltip = 'Verified to be supportive, helpful people';
-
-  return <TooltipContainer id="verified-team-tooltip" type="info" tooltip={tooltip} target={<img className="verified" src={image} alt="✓" />} />;
-};
-
-
 export class WhitelistedDomainIcon extends React.Component {
   constructor(props) {
     super(props);
